fix int in db

+++ contrib/migrations/002_add_wallets.sql
@@ -1,12 +1,12 @@
 CREATE TABLE wallets
 (
-    user_id          INTEGER UNIQUE REFERENCES users(user_id),
-    microdollars     INTEGER NOT NULL
+    user_id          BIGINT UNIQUE REFERENCES users(user_id),
+    microdollars     BIGINT NOT NULL
 );
 
 CREATE TABLE wallets_history
 (
-    user_id          INTEGER REFERENCES users(user_id),
-    microdollars     INTEGER NOT NULL,
+    user_id          BIGINT REFERENCES users(user_id),
+    microdollars     BIGINT NOT NULL,
     created_at       timestamptz NOT NULL DEFAULT NOW()
 );
